refactor(button): add styles and types button files
