+++ src/components/button/index.ts
@@ -1,1 +1,3 @@
 export * from './button';
+export * from './button.styles';
+export * from './button.types';
